back-end/agent/requirements.txt
fastapi
uvicorn[standard]
google-adk
google-adk[a2a]
google-adk[Runner]
google-genai
requests
